Refactor validator fixture harnesses
